requirements.txt
Flask==2.2.3
pandas==1.5.3
gunicorn==20.1.0
opencv-python-headless==4.4.0.44
numpy==1.24.2
Pillow==9.4.0
tensorflow==2.11.0
Werkzeug==2.2.2
pandastable
flask_cors